Bug 1217970 - out-of-line ScrollbarStyles::IsSmoothScroll; r=bz
IsSmoothScroll depends on the exact values of dom::ScrollBehavior, which
is defined in Window.webidl.  Therefore, to define those values for
IsSmoothScroll, we would need to include WindowBinding.h.  That file,
however, drags in many different dependencies to many different
places (e.g. IPDL generated protocols), which is undesirable.

To work around this, define the function out-of-line so that only
ScrollbarStyles.cpp needs to include WindowBinding.h.  Doing this cuts
down files that depend on WindowBinding.h by ~25%.