part 1a - refactor BluetoothDaemonSetupInterface; r=btian
This is the first in a series of patches to reduce code duplication in
bluedroid/.  We have the common pattern:

  create DaemonSocketPDU
  PackPDU, check for failure
  Send PDU, check for failure
  drop reference to newly-created socket
  return

in every single interface implementation.  While we can't completely
factor the code out into a common place (since everything needs to call
Send()), we can at least factor it out into common code per-interface.