Bug 1239421 - part 1 - create nsIKeyObject{,Factory} instances directly; r=keeler
There's no need for things in security/manager/ssl/ to go through XPCOM
to create instances of nsIKeyObject{,Factory}.

Furthermore, the code in nsNTLMAuthModule.cpp was creating nsIKeyObject
objects only to blow them away later with calls to
nsIKeyObjectFactory::KeyFromString.  That's useless work, so let's just
get rid of it.