Bug 1369789 - turn off -Wmaybe-uninitialized for IPDL code; r=jld
As the comment for the accompanying change says, GCC complains about a
handful of instances of this pattern:

    ActorHandle handle__;
    ...
    // initializes handle__ if successful
    if ((!(Read((&(handle__)), (&(msg__)), (&(iter__)))))) {
        FatalError("Error deserializing 'ActorHandle'");
        return MsgValueError;
    }
    // more code...
    RegisterID(actor, (handle__).mId);

saying that mId is potentially uninitialized.  What's curious is that it
doesn't complain about the scores of other instances; I do not know what
makes it complain about the particular instances that it does.  Let's
just turn the warning off.