Bug 1383911 - part 1 - rationalize the allprotocols list; r=billm
The allprotocols list suggests that it's, well, a list of all the
protocols.  What it actually is is a list of all the ${protocol}MsgStart
constants, along with some unused ${protocol}MsgStartChild constants.
We can remove the latter.  For the former, we can inject the MsgStart
suffix at another point, therefore making allprotocols's name more
closely reflect its intent.