provide a generic framework for dumping leaked things at shutdown; r=mccr8
The URL code has a mechanism for dumping the addresses of leaked URLs at
shutdown, so people have a clue where to start looking for leaks.  We'd
like to extend this behavior to other classes, so let's generalize the
code for doing so.  We'll have less boilerplate to write this way.