Bug 1435766 - part 2 - reimplement nsDirectoryService Mac getters in terms of GetSpecialSystemDirectory; r=spohl
This change makes the nsDirectoryService implementation for Mac more
consistent with the Windows and generic Unix implementations.  The
change also opens up the possibility of using GetSpecialSystemDirectory
on Mac for Mac-specific things elsewhere in the codebase and not having
things fall over.  We need to add a bunch of SystemDirectories constants
to accommodate all the atoms required by the directory service.