Bug 1428158 - part 2 - make a modifiable version of OS_LIBS; r=build-peer
The clang static analysis plugin is, for reasons, built as a target
library, even though it's really a host library.  This decision
necessitates various hacks when building the clang plugin, most of which
are confined to the clang-plugin build directory.

For target code on Android, however, we need to link in several target
libraries, most notably the C++ standard library and various supporting
bits.  Trying to link a host library while using these target-specific
libraries does not work very well.

Thus this hack.  We need to remove the target-specific libraries when
linking the clang plugin, and only the clang plugin.  But we can't
modify OS_LIBS, as moz.build knows about OS_LIBS and will complain if
Makefile.in files touch OS_LIBS.  So instead we need a level of
indirection, a variable whose sole purpose is to wrap OS_LIBS so the
clang plugin Makefile.in can remove various bits from it.