Bug 1428158 - part 5 - deal with libstdc++ incompatibilities; r=build-peer
Loading the clang plugin was failing completely without any useful error
messages.  LD_DEBUG showed me that it was failing to resolve several
symbols, presumably because said methods aren't defined in the
libstdc++.a files that come with the clang packaged up on infra (because
we built the whole thing with an older compiler there?).  This is going
to be super-annoying for people who want to make the plugin work
locally, but we can cross that bridge when we come to it; maybe by using
the clang on their system instead of one downloaded from `mach artifact
toolchain` will magically resolve this.

Anyway, some basic_string methods were taken care of by using the old
libstdc++ ABI.  Out-of-line definitions for error_category's virtual
methods were also necessary.

These steps are really similar to what we need for
--enable-stdcxx-compat, but I was hesistant to go down that rabbit hole
on Android, since that could be a whole heap of fun to deal with.  I'd
rather get everything working, and then sort of what the best way from
"working locally" to "working on infra" is.