Bug 1459034 - part 2 - devirtualize nsIFrame::GetContentOffsetsFromPointExternal; r=dholbert
The default implementation of this function just forwards to
GetContentOffsetsFromPoint, and there are no subclasses that override
this function.  I believe the "External" prefix hearkens back to a day
when bits of layout lived in a separate library from everything else,
but that day is long gone.  Both reasons are sufficient motivation to
devirtualize this function.